add x86_pc auto update-ˇ

+++ Client/x86_PC/environment.py
@@ -28,7 +28,10 @@ os.system("pip3 install influxdb")
 
 # supervisor install
 os.system("apt-get install -y supervisor")
-os.system("cp /home/ubuntu/School_Monitor_System/Client/x86_PC/client.conf /etc/supervisor/conf.d")
+
+if (not os.path.exists("/etc/supervisor/conf.d/client.conf")):
+    os.system("cp /home/ubuntu/School_Monitor_System/Client/x86_PC/client.conf /etc/supervisor/conf.d")
+    os.system("cp /home/ubuntu/School_Monitor_System/Client/x86_PC/client.conf.bak /home/ubuntu/School_Monitor_System/Client/x86_PC/client.conf")
 os.system("service supervisor restart")
 while True:
     if (len(str(check_output(["pidof","python3"]).decode("utf-8")).split("\n")[0].split(" ")) >= 3):
